Remove lab complited

+++ VirtualLab/Application/LabManager.cs
@@ -70,7 +70,7 @@ public class LabManager : ILabManager
     public async Task<Result> End(Guid labId, Guid userId)
     {
         var getUserLab = await _userLabProvider.GetUserLab(userId, labId);
-        if (getUserLab.TryGetValue(out var userLabInfo))
+        if (!getUserLab.TryGetValue(out var userLabInfo))
         {
             _log.Error($"Not find lab with {labId} for user with id {userId}"); // так бы везде))
             return Result.Fail(getUserLab.Errors);
@@ -83,7 +83,8 @@ public class LabManager : ILabManager
             return Result.Fail($"error {getUserLab.Errors}");
         }
 
-        await _stands.Delete(config);
+        var deletedConfig = await _stands.Delete(config);
+        if (deletedConfig.IsFailed) return deletedConfig;
 
         return await _virtualMachineDataHandler.DeleteAllByUserLabId(userLabInfo.Id);
     }
